requirements.txt
python-telegram-bot
requests
gql[all]
aiohttp
pydantic
pydantic-settings
